+++ src/test/java/test/TestSampleCode.java
@@ -53,7 +53,7 @@ public class TestSampleCode {
       boolean isIncoming = transfer.getIsIncoming();
       BigInteger amount = transfer.getAmount();
       int accountIdx = transfer.getAccountIndex();
-      Integer height = transfer.getTx().getHeight();  // will be null if unconfirmed
+      Long height = transfer.getTx().getHeight();  // will be null if unconfirmed
     }
     
     // get incoming transfers to account 0
@@ -62,7 +62,7 @@ public class TestSampleCode {
       assertTrue(transfer.getIsIncoming());
       assertEquals(0, (int) transfer.getAccountIndex());
       BigInteger amount = transfer.getAmount();
-      Integer height = transfer.getTx().getHeight();  // will be null if unconfirmed
+      Long height = transfer.getTx().getHeight();  // will be null if unconfirmed
     }
 
     // send to an address from account 0
@@ -100,7 +100,7 @@ public class TestSampleCode {
     //MoneroDaemon daemon = new MoneroDaemonRpc("http://localhost:38081", "admin", "password");
     
     // get daemon info
-    int height = daemon.getHeight();                  // e.g. 1523651
+    long height = daemon.getHeight();                 // e.g. 1523651
     BigInteger feeEstimate = daemon.getFeeEstimate(); // e.g. 750000
     
     // get last block's header
@@ -108,11 +108,11 @@ public class TestSampleCode {
     long lastBlockSize = lastBlockHeader.getSize();
     
     // get first 100 blocks as a binary request
-    List<MoneroBlock> blocks = daemon.getBlocksByRange(0, 100);
+    List<MoneroBlock> blocks = daemon.getBlocksByRange(0l, 100l);
     
     // get block info
     for (MoneroBlock block : blocks) {
-      int blockHeight = block.getHeight();
+      long blockHeight = block.getHeight();
       String blockId = block.getId();
       List<MoneroTx> txs = block.getTxs();
